Avancé en las funcionalidades de gestión de usuarios y registros, desde el admin

+++ services/register.py
@@ -15,8 +15,14 @@ class RegisterService():
 
         return result
 
-    def create_register(self, register: Register) -> None:
-        new_register = RegisterModel(**register.dict())
+    def create_register(self, id: int, register_name: str, meassure: float, meassure_unit: str) -> None:
+        
+        register = self.get_register(id=id)
+
+        if register:
+            raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="El id ya pertenece a un registro")
+        
+        new_register = RegisterModel(id=id, register_name=register_name, meassure=meassure, meassure_unit=meassure_unit)
 
         self.db.add(new_register)
         self.db.commit()
@@ -43,6 +49,9 @@ class RegisterService():
     def delete_register(self, id: int) -> None:
         result = self.db.query(RegisterModel).filter(RegisterModel.id == id).first()
 
+        if not result:
+            raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="No se encontró el registro")
+        
         self.db.delete(result)
         self.db.commit()
